+++ Sources/TidalAPIInterchangeKit/API/Models/TidalAlbumAttributes.swift
@@ -18,13 +18,13 @@ public struct TidalAlbumAttributes: Codable, Equatable, Hashable, Sendable {
 
     public let albumType: String
 
-    public let barcodeId: String
+    public let barcodeID: String
 
     public let copyright: TidalCopyright?
 
-    public let duration: String
+    public let iso8601Duration: String
 
-    public let explicit: Bool
+    public let isExplicit: Bool
 
     public let externalLinks: [TidalExternalLink]?
 
@@ -36,37 +36,65 @@ public struct TidalAlbumAttributes: Codable, Equatable, Hashable, Sendable {
 
     public let popularity: Double
 
-    public let releaseDate: String?
+    public let iso8601ReleaseDate: String?
 
     public let version: String?
 
     public init(title: String,
                 accessType: String? = nil,
                 albumType: String,
-                barcodeId: String,
+                barcodeID: String,
                 copyright: TidalCopyright? = nil,
-                duration: String,
-                explicit: Bool,
+                iso8601Duration: String,
+                isExplicit: Bool,
                 externalLinks: [TidalExternalLink]? = nil,
                 mediaTags: [String],
                 numberOfItems: Int32,
                 numberOfVolumes: Int32,
                 popularity: Double,
-                releaseDate: String? = nil,
+                iso8601ReleaseDate: String? = nil,
                 version: String? = nil) {
         self.title = title
         self.accessType = accessType
         self.albumType = albumType
-        self.barcodeId = barcodeId
+        self.barcodeID = barcodeID
         self.copyright = copyright
-        self.duration = duration
-        self.explicit = explicit
+        self.iso8601Duration = iso8601Duration
+        self.isExplicit = isExplicit
         self.externalLinks = externalLinks
         self.mediaTags = mediaTags
         self.numberOfItems = numberOfItems
         self.numberOfVolumes = numberOfVolumes
         self.popularity = popularity
-        self.releaseDate = releaseDate
+        self.iso8601ReleaseDate = iso8601ReleaseDate
         self.version = version
     }
+
+    enum CodingKeys: String, CodingKey {
+        case title
+        case accessType
+        case albumType
+        case barcodeID = "barcodeId"
+        case copyright
+        case iso8601Duration = "duration"
+        case isExplicit = "explicit"
+        case externalLinks
+        case mediaTags
+        case numberOfItems
+        case numberOfVolumes
+        case popularity
+        case iso8601ReleaseDate = "releaseDate"
+        case version
+    }
+}
+
+public extension TidalAlbumAttributes {
+
+    var releaseDate: Date? {
+        guard let iso8601ReleaseDate else { return nil }
+
+        return try? Date.ISO8601FormatStyle.dateOnlyStyle.parse(iso8601ReleaseDate)
+    }
+
+    // TODO: add getter to convert iso8601Duration to TimeInterval
 }