+++ Sources/TidalAPIInterchangeKit/API/Models/TidalVideoAttributes.swift
@@ -14,9 +14,9 @@ public struct TidalVideoAttributes: Codable, Equatable, Hashable, Sendable {
 
     public let copyright: TidalCopyright?
 
-    public let duration: String
+    public let iso8601Duration: String
 
-    public let explicit: Bool
+    public let isExplicit: Bool
 
     public let externalLinks: [TidalExternalLink]?
 
@@ -24,30 +24,52 @@ public struct TidalVideoAttributes: Codable, Equatable, Hashable, Sendable {
 
     public let popularity: Double
 
-    // Release date (ISO-8601)
-    public let releaseDate: String?
+    public let iso8601ReleaseDate: String?
 
     public let title: String
 
     public let version: String?
 
     public init(copyright: TidalCopyright? = nil,
-                duration: String,
-                explicit: Bool,
+                iso8601Duration: String,
+                isExplicit: Bool,
                 externalLinks: [TidalExternalLink]? = nil,
                 isrc: String,
                 popularity: Double,
-                releaseDate: String? = nil,
+                iso8601ReleaseDate: String? = nil,
                 title: String,
                 version: String? = nil) {
         self.copyright = copyright
-        self.duration = duration
-        self.explicit = explicit
+        self.iso8601Duration = iso8601Duration
+        self.isExplicit = isExplicit
         self.externalLinks = externalLinks
         self.isrc = isrc
         self.popularity = popularity
-        self.releaseDate = releaseDate
+        self.iso8601ReleaseDate = iso8601ReleaseDate
         self.title = title
         self.version = version
     }
+
+    enum CodingKeys: String, CodingKey {
+        case copyright
+        case iso8601Duration = "duration"
+        case isExplicit = "explicit"
+        case externalLinks
+        case isrc
+        case popularity
+        case iso8601ReleaseDate = "releaseDate"
+        case title
+        case version
+    }
+}
+
+public extension TidalVideoAttributes {
+
+    var releaseDate: Date? {
+        guard let iso8601ReleaseDate else { return nil }
+
+        return try? Date.ISO8601FormatStyle.dateOnlyStyle.parse(iso8601ReleaseDate)
+    }
+
+    // TODO: add getter to convert iso8601Duration to TimeInterval
 }